Refactor tool invocation code and refine prompt template
